adjusted value type in customer table from integer to deciaml for accoutnt credit.

+++ db/setup/schema.sql
@@ -20,7 +20,7 @@ CREATE TABLE customers(
   state text,
   postal_code text,
   phone text,
-  account_credit integer
+  account_credit decimal
 );
